requirements.txt
streamlit==1.46.1
pandas==2.2.3
openpyxl==3.1.5
plotly==5.24.1
numpy==2.2.0
altair==5.5.0
python-dotenv==1.0.0
langchain-openai==0.2.9
langchain-core==0.3.28
langgraph==0.2.45
openai==1.58.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
supabase==2.8.1 